backend/requirements.txt
fastapi
uvicorn[standard]
pydantic

faiss-cpu
duckdb
pandas
numpy

# For LLMs
openai
google-genai

# Mongo
pymongo
psycopg2

# Embedding / ML
python-dotenv
fastembed 

# JWT
python-jose[cryptography]
python-multipart
bcrypt==4.0.1